+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -1,6 +1,5 @@
-package roomescape.controller;
+package roomescape.reservation.controller;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
@@ -13,7 +12,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.entity.Reservation;
+import roomescape.reservation.entity.Reservation;
 
 @RestController
 @RequestMapping("/reservations")
